feat: implement the DECR command.
As part of this work, we've added the value_entry_t type to the hashtable_entry_t so we can store key, value and associated metadata.

This enables us to store the values underlying encoding. For example, when a value is an integer we use the encoding VALUE_ENTRY_TYPE_INT. This ensures we can operate on the underlying value in the best way possible according to it's encoding. if the value is a string, we know we can't do DECR and INCR operations, if the value is an int, we can grab whatever is ptr is convert to int and do operations accordingly.